Improve makefile and setup scripts for easier installation.

+++ vs_code/setup.sh
@@ -1,28 +1,10 @@
 #!/bin/bash
 
-red="$(tput setaf 1)"
-green="$(tput setaf 2)"
-cyan="$(tput setaf 6)"
-reset="$(tput sgr0)"
-
-newline='
-'
-
-log() {
-	printf "${newline}${green}INFO: $1${reset}"
-}
-
-error() {
-	printf "${newline}${red}ERROR: $1${reset}"
-}
-
 initialize() {
-	printf "${newline}"
-	read -r -p "${cyan}Setup VS Code config? ${reset}[y/N] " answer
+	read -r -p "Include vs code? [y/N] " answer
 	if [ "$answer" != y ] && [ "$answer" != Y ]; then
-		error "VS Code setup declined..."
-		error "Now exiting."
-		printf "${newline}${newline}"
+		echo -e "\nExcluding vs code..."
+		echo -e "Done.\n"
 		exit
 	else
 		setup
@@ -30,17 +12,21 @@ initialize() {
 }
 
 setup() {
-	log "Setting up VS Code config..."
+	echo -e "\nSetting up vs code config..."
 
-	if [ -f $HOME/Library/Application\ Support/Code/User/settings.json ]; then
-		rm $HOME/Library/Application\ Support/Code/User/settings.json
+	if [[ ! $(ls /Applications/ | grep -i visual\ studio\ code) ]]; then
+		echo -e "\nCould not find existing vs code installation."
+		echo "Please install it manually."
+	else
+		if [ -f $HOME/Library/Application\ Support/Code/User/settings.json ]; then
+			rm $HOME/Library/Application\ Support/Code/User/settings.json
+		fi
+
+		echo -e "\nSymlinking settings.json..."
+		ln -sf "$(pwd)/settings.json" $HOME/Library/Application\ Support/Code/User/settings.json
 	fi
 
-	log "Symlinking settings.json..."
-	ln -sf "$(pwd)/settings.json" $HOME/Library/Application\ Support/Code/User/settings.json
-
-	log "VS Code setup complete."
-	printf "${newline}${newline}"
+	echo -e "Done.\n"
 }
 
 initialize
